requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
av==17.0.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
comm==0.2.3
confection==1.3.3
cryptography==46.0.7
cymem==2.0.13
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.2.1
faiss-cpu==1.13.2
filelock==3.25.2
fsspec==2026.3.0
google-auth==2.49.1
google-genai==1.71.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.0
idna==3.11
igraph==1.0.0
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
leidenalg==0.11.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.15
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
openai==2.30.0
packaging==26.0
pandas==3.0.2
parso==0.8.6
pillow==12.2.0
platformdirs==4.9.4
preshed==3.0.13
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
pyzmq==27.1.0
qwen-vl-utils==0.0.14
regex==2026.4.4
requests==2.33.1
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.4
texttable==1.7.0
thinc==8.3.13
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0+cu126
torchvision==0.26.0+cu126
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.5.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wasabi==1.1.3
wcwidth==0.6.0
weasel==1.0.0
websockets==16.0
wheel==0.46.3
wrapt==2.1.2
